requirements.txt
Flask
werkzeug
openai-whisper
